adding new item to order works, trying to add item to order that already has item doesnt error out

+++ server/db/models/item_in_order.js
@@ -6,7 +6,7 @@ const ItemInOrder = db.define('item-in-order', {
     type: Sequelize.SMALLINT
   },
   purchaseTotal: {
-    type: Sequelize.DECIMAL
+    type: Sequelize.SMALLINT
   }
 })
